重构：重命名Controller并优d相关类

+++ module/src/main/java/red/mlz/entity/Card.java
@@ -4,12 +4,12 @@ import lombok.Data;
 
 @Data
 public class Card {
-    private Integer id;
+    private Long id;
     private String coverImages;
     private String name;
     private Float price;
     private String introduction;
     private Long createTime;
     private Long updateTime;
-    private int isDeleted;
+    private Integer isDeleted;
 }
